+++ lib/message.dart
@@ -5,14 +5,30 @@ part 'message.g.dart';
 
 @JsonSerializable()
 class Message extends Object with _$MessageSerializerMixin {
-	Message(this.time,this.count,this.elements,this.sender);
+	Message(this.time,this.count,this.elements,this.sender,this.isSelf,this.conversationType,this.conversationId);
 
 	int time;
 	int count;
 	String sender;
+	bool isSelf;
 	List<MessageElement> elements;
+	String conversationType;
+	String conversationId;
 
 	factory Message.fromJson(Map<String, dynamic> json) => _$MessageFromJson(json);
 
-	static Message createFromJson(dynamic json) => _$MessageFromJson(json);
+	static Message createFromJson(dynamic json){
+		return new Message(
+				json['time'] as int,
+				json['count'] as int,
+				(json['elements'] as List)
+						?.map((e) => e == null
+						? null
+						: MessageElement.createFromJson(e))
+						?.toList(),
+				json['sender'] as String,
+				json['isSelf'] as bool,
+				json['conversationType'] as String,
+				json['conversationId'] as String,);
+	}
 }